minor test & testkit refactoring
